+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Method `count_parameters` to `TModule`.
 - Option `padding_idx` to `indices_to_onehot` function.
-- Function `dict_list_to_list_dict` to collections utils.
+- Functions `dict_list_to_list_dict`, `flat_list`, `unflat_dict_of_dict` to collections utils.
 - Class `PickleDataset` and function `pack_to_pickle` to utils.
 - Class `ResampleNearest` and function `resample_nearest`.
 - Class `TransformDrop` and function `transform_drop`.